requirements.txt
python-can
numpy
scipy
pyqt5
pyqtgraph
